+++ src/rules/function-whitelist/index.js
@@ -9,23 +9,23 @@ import {
   validateOptions,
 } from "../../utils"
 
-export const ruleName = "function-blacklist"
+export const ruleName = "function-whitelist"
 
 export const messages = ruleMessages(ruleName, {
   rejected: (name) => `Unexpected function "${name}"`,
 })
 
-export default function (blacklist) {
+export default function (whitelist) {
   return (root, result) => {
     const validOptions = validateOptions(result, ruleName, {
-      actual: blacklist,
+      actual: whitelist,
       possible: [isString],
     })
     if (!validOptions) { return }
     root.walkDecls(decl => {
       const { value } = decl
       valueParser(value).walk(function (node) {
-        if (node.type === "function" && blacklist.indexOf(vendor.unprefixed(node.value)) !== -1) {
+        if (node.type === "function" && whitelist.indexOf(vendor.unprefixed(node.value)) === -1) {
           report({
             message: messages.rejected(node.value),
             node: decl,